BUILD-1.33 keep persisted progress strongly typed

+++ app/lib/core/data/persistent_library.dart
@@ -253,7 +253,7 @@ DownloadJob? _downloadFromJson(Object? value) {
       id: id,
       source: source,
       state: DownloadState.values.byName(state),
-      progress: progress is num ? progress.toDouble().clamp(0, 1) : 0,
+      progress: progress is num ? progress.toDouble().clamp(0, 1).toDouble() : 0,
       localPath: localPath is String ? localPath : null,
       error: error is String ? error : null,
     );
